Program is correctly created
Code was added to ensure that:
The file names created parsed by the program excluded the ".txt" distinction. This is to prevent duplicate feature files from being created.

It All ensure that a program is creating creating by accurately assembled the beginning, middle, and end of the code by read input from feature files and program objects.